+++ src/reader.cpp
@@ -2743,7 +2743,7 @@ const wchar_t *reader_readline(int nchars) {
                 if (data->search_mode) {
                     data->search_mode = NO_SEARCH;
 
-                    if (data->token_history_pos == -1) {
+                    if (data->token_history_pos == (size_t)-1) {
                         // history_reset();
                         data->history_search.go_to_end();
                         reader_set_buffer(data->search_buff, data->search_buff.size());
